:recycle:refactor: refactor bookshelves page 컴포넌트, 스타일 분리 Related to: #68
